Wrote full test case for signup interactor
